Rename supply_info.rs to asset_record.rs
