+++ src/asset_record.rs
