+++ src/lib.rs
@@ -18,6 +18,7 @@
 
 mod action;
 mod address;
+pub mod asset_record;
 pub mod builder;
 pub mod bundle;
 pub mod circuit;
@@ -29,7 +30,6 @@ pub mod note;
 pub mod orchard_flavor;
 pub mod primitives;
 mod spec;
-pub mod supply_info;
 pub mod tree;
 pub mod value;
 pub mod zip32;
